Add patch series support to sync command
Add patch series support to sync command so that users would be able to
sync their patch series.

[changelog]
added: patch series support to the sync command

<!-- ps-id: 198d2e18-8144-4ef5-99e8-2a1b07a306b6 -->
