+++ src/cli.rs
@@ -58,7 +58,7 @@ pub struct ShowCmdOpts {
 
 #[derive(Debug, Args)]
 pub struct SyncCmdOpts {
-    pub patch_index: usize,
+    pub patch_index_or_range: String,
     /// Use the provided branch name instead of generating one
     #[arg(short = 'n')]
     pub branch_name: Option<String>,
