+++ src/main.rs
@@ -49,7 +49,9 @@ fn main() {
             commands::batch_request_review::batch_request_review(opts.patch_index, cli.color)
         }
         cli::Command::Show(opts) => commands::show::show(opts.patch_index),
-        cli::Command::Sync(opts) => commands::sync::sync(opts.patch_index, opts.branch_name),
+        cli::Command::Sync(opts) => {
+            commands::sync::sync(opts.patch_index_or_range, opts.branch_name)
+        }
         cli::Command::Isolate(opts) => {
             commands::isolate::isolate(opts.start_patch_index, opts.end_patch_index, cli.color)
         }
