+++ src/ps/public/request_review.rs
@@ -121,7 +121,7 @@ pub fn request_review(
 
     // sync patch up to remote
     let (patch_upstream_branch_name, _patch_upstream_branch_remote_name) =
-        ps::public::sync::sync(patch_index, given_branch_name)
+        ps::public::sync::sync(patch_index, None, given_branch_name)
             .map_err(RequestReviewError::SyncFailed)?;
 
     // execute post sync hook
